avoid warnings in Google Chrome console of deprecated APIs
Resolves issue https://github.com/jakutis/httpinvoke/issues/12 that
was initiated by Zak Angelle <zak@nuclearhelix.com> (https://github.com/zakangelle)
and commented by Sebastian Ware <sebastian@urbantalk.se> (https://github.com/jhsware).

zakangele:
According to the ProgressEvent API, the XMLHttpRequest
ProgressEvent.position property is now deprecated, and using it shows a
warning in the Chrome console.
Is there any reason (e.g. backwards compatibility) that these references
shouldn't be removed?

jhsware:
@zakangelle The code
    progressEvent.loaded || progressEvent.position || 0
means that it should use progressEvent.loaded if available and only use
progressEvent.position if it doesn't find the former. I got the same
errors and looked in the code. Removing this like you have done breaks
backward compatibility. However it seems that there is a case early in
the call when progressEvent.loaded is undefined and it would be nice if
it didn't throw this warning in those cases. But you can't just remove
progressEvent.position. You need to do a proper check to figure out
which one you want to use.